Implement name functionality, str & type hints
Implement the type hints for mypy in the whole "reservation.py"
document.

Implement the dunder str method.

Implement the functionality of the Reservation class related to the
information of the name of a Reservation instance. In this case, the
getter and setter of the name attribute and a method to clean the user
input for it.

Fix typos and dosctrings.