Implement date functionality and refactor str
Refactor dunder str method of the Reservation class.

Refactor and create the docstrings of getters, setters and methods of
name and date properties of the Reservation class.

Implement the functionality of the Reservation class related to the
information of the date of a Reservation instance. In this case,
the getter and setter of the date attribute and a method to clean the
user input for it.