Fix bug: RecursionError
After Reservation class attibutes, getters and setters nomenclature
were changed to refactor their privacy, Python raised a RecursionError
when a user tried to create a reservation because of attributes,
getters and setters had the same name then. So I modified the
attributes, getters and setters nomenclature on the one hand and the
nomenclature of their internal values on the other hand to fix this bug.