Refactor "update _database" Method
Improved docstrings and comments for better clarity.

Renamed variables for consistent and meaningful naming.

Refactored the conversion of the new reservation dictionary.

Implement time-based sorting for the reservations database, after the
date-based sorting.

Enhanced reservation enumeration logic.

Removed redundant confirmation document exportation, as it is
already handled in the "create_reservation" method and was not
appropiate to perform in this method.