Fix bug and refactor update and delete methods
Change the name of delete_reservation method to cancel_reservation and
change the name in lines where this method is called.

Fix bug in the cancel_reservation: The operator "is not" wasn't correct
in this case where i wanted to compare strings.

Refactor update_reservation method and complete its docstring.

Modify de exits functions in the main method to better functionality.