Move the validation methods out of class
To meet project requirements, I've moved the
validation methods out of the class Reservation
to perform test on them later. Also, I've realized
that these methods do not have a specific
relationship with the class and therefore can be
moved outside of it.

Fix some typo.

Refactor main method docstring.